turned getRand from double to int, and removed it

+++ PTPWRd/dep/ptpd_dep.h
@@ -171,7 +171,7 @@ void displayStats(RunTimeOpts *rtOpts, PtpClock *ptpClock);
 Boolean nanoSleep(TimeInternal*);
 void getTime(TimeInternal*);
 void setTime(TimeInternal*);
-double getRand();
+int getRand();
 Boolean adjFreq(Integer32);
 /** \}*/
